upds

+++ lvl_3_minishell/main.c
@@ -6,13 +6,13 @@
 /*   By: nnuno-ca <nnuno-ca@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/10 02:02:08 by nnuno-ca          #+#    #+#             */
-/*   Updated: 2023/02/19 14:52:14 by nnuno-ca         ###   ########.fr       */
+/*   Updated: 2023/02/19 22:54:16 by nnuno-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	g_exit_status = 0;
+long long	g_exit_status = 0;
 
 static char	*get_input(void)
 {
