+++ lvl_3_minishell/srcs/config_signals.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   config_signals.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: roramos <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: nnuno-ca <nnuno-ca@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/29 16:54:55 by roramos           #+#    #+#             */
-/*   Updated: 2023/02/11 16:24:08 by roramos          ###   ########.fr       */
+/*   Updated: 2023/02/19 22:53:04 by nnuno-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-extern int	g_exit_status;
+extern long long	g_exit_status;
 
 void	child_signals(int signum)
 {
